requirements.txt
django
django-sphinx
django-tagging
